refactor: extract AccountService from AuthenticationService
AuthenticationService carried four dependencies, three of them optional,
so `new AuthenticationService(userRepository)` type-checked while half
its methods were unusable. apps/admin and the two grant scripts all
constructed it that way.

Split along the dependency line. AuthenticationService keeps login,
grantAccess, grantAdmin, and changePassword — everything needing only
the user store — and drops to a single required constructor argument.
AccountService takes register, updateEmail, requestPasswordReset,
resetPassword, and validateResetToken, all of which need the reset-token
store and a way to send mail.

Pure move: no behaviour change, and AccountService keeps the optional
constructor args for now. Making them required is the next commit.

The test file splits along the same seam. auth.tsx and profile.tsx pick
up the second service; private.tsx and apps/admin are untouched, and
admin now constructs a service whose every method works.
